+++ R/transformer_deberta_v2.R
@@ -12,10 +12,6 @@
 #'vocabulary.
 #'@param vocab_size \code{int} Size of the vocabulary.
 #'@param do_lower_case \code{bool} If \code{TRUE} all characters are transformed to lower case.
-#'@param add_prefix_space \code{bool} \code{TRUE} if an additional space should be insert
-#'to the leading words.
-#'@param trim_offsets \code{bool} If \code{TRUE} post processing trims offsets
-#'to avoid including whitespaces.
 #'@param max_position_embeddings \code{int} Number of maximal position embeddings. This parameter
 #'also determines the maximum length of a sequence which can be processed with the model.
 #'@param hidden_size \code{int} Number of neurons in each layer. This parameter determines the
@@ -49,9 +45,9 @@
 #'and the vocabulary of the new model are saved on disk.
 #'@note To train the model, pass the directory of the model to the function
 #'\link{train_tune_deberta_v2_model}.
-#'@note For this model a SentencePiece tokenizer is created which does not rely
-#'on the corresponding python library sentencepiece. Instead an implementation
-#'of the tokenizer library.
+#'@note For this model a WordPiece tokenizer is created. The standard implementation
+#'of DeBERTa version 2 from HuggingFace uses a SentencePiece tokenizer. Thus, please
+#'use \code{AutoTokenizer} from the 'transformers' library to use this model.
 #'
 #'@references
 #'He, P., Liu, X., Gao, J. & Chen, W. (2020). DeBERTa: Decoding-enhanced BERT
@@ -70,8 +66,6 @@ create_deberta_v2_model<-function(
     model_dir,
     vocab_raw_texts=NULL,
     vocab_size=128100,
-    add_prefix_space = FALSE,
-    trim_offsets=TRUE,
     do_lower_case=FALSE,
     max_position_embeddings=512,
     hidden_size=1536,
@@ -183,20 +177,41 @@ create_deberta_v2_model<-function(
               "Creating Tokenizer Draft"))
   }
 
-  tok_new<-tok$SentencePieceUnigramTokenizer()
-  tok_new$normalizer<-tok$normalizers$BertNormalizer(
+  #tok_new<-tok$SentencePieceUnigramTokenizer()
+  #tok_new$normalizer<-tok$normalizers$BertNormalizer(
+  #  clean_text = TRUE,
+  #  handle_chinese_chars = TRUE,
+  #  strip_accents = do_lower_case,
+  #  lowercase = do_lower_case
+  #)
+  #tok_new$post_processor<-tok$processors$RobertaProcessing(
+  #  sep=reticulate::tuple(list("[SEP]",as.integer(1))),
+  #  cls=reticulate::tuple(list("[CLS]",as.integer(0))),
+  #  trim_offsets=trim_offsets,
+  #  add_prefix_space = add_prefix_space
+  #)
+  #tok_new$enable_padding(pad_token = "[PAD]")
+
+  special_tokens=c("[CLS]","[SEP]","[PAD]","[UNK]","[MASK]")
+  tok_new<-tok$Tokenizer(tok$models$WordPiece())
+  tok_new$normalizer=tok$normalizers$BertNormalizer(
+    lowercase=do_lower_case,
     clean_text = TRUE,
     handle_chinese_chars = TRUE,
-    strip_accents = do_lower_case,
-    lowercase = do_lower_case
-  )
-  tok_new$post_processor<-tok$processors$RobertaProcessing(
+    strip_accents = do_lower_case)
+  tok_new$pre_tokenizer=tok$pre_tokenizers$BertPreTokenizer()
+  tok_new$post_processor<-tok$processors$BertProcessing(
     sep=reticulate::tuple(list("[SEP]",as.integer(1))),
-    cls=reticulate::tuple(list("[CLS]",as.integer(0))),
-    trim_offsets=trim_offsets,
-    add_prefix_space = add_prefix_space
+    cls=reticulate::tuple(list("[CLS]",as.integer(0)))
   )
-  tok_new$enable_padding(pad_token = "[PAD]")
+
+  tok_new$decode=tok$decoders$WordPiece()
+
+  trainer<-tok$trainers$WordPieceTrainer(
+    vocab_size=as.integer(vocab_size),
+    special_tokens = special_tokens,
+    show_progress=trace)
+
 
   update_aifeducation_progress_bar(value = 3,
                                    total = pgr_max,
@@ -217,15 +232,20 @@ create_deberta_v2_model<-function(
       shiny_app_active=TRUE
     }
   }
-  tok_new$train_from_iterator(py$batch_iterator(batch_size = as.integer(2),
+  #tok_new$train_from_iterator(py$batch_iterator(batch_size = as.integer(2),
+  #                                              dataset=raw_text_dataset,
+  #                                              report_to_shiny_app=shiny_app_active),
+  #                            #trainer=trainer,
+  #                            length=length(raw_text_dataset),
+  #                            vocab_size = as.integer(vocab_size),
+  #                            special_tokens=c("[CLS]","[SEP]","[PAD]","[UNK]","[MASK]"),
+  #                            unk_token="[UNK]"
+  #                            )
+  tok_new$train_from_iterator(py$batch_iterator(batch_size = as.integer(200),
                                                 dataset=raw_text_dataset,
                                                 report_to_shiny_app=shiny_app_active),
-                              #trainer=trainer,
-                              length=length(raw_text_dataset),
-                              vocab_size = as.integer(vocab_size),
-                              special_tokens=c("[CLS]","[SEP]","[PAD]","[UNK]","[MASK]"),
-                              unk_token="[UNK]"
-                              )
+                              trainer=trainer,
+                              length=length(raw_text_dataset))
 
 
   if(trace==TRUE){
@@ -249,7 +269,11 @@ create_deberta_v2_model<-function(
     dir.create(model_dir)
   }
 
-  tok_new$save_model(model_dir)
+
+  write(c(special_tokens,names(tok_new$get_vocab())),
+        file=paste0(model_dir,"/","vocab.txt"))
+
+  #tok_new$save_model(model_dir)
 
   update_aifeducation_progress_bar(value = 5,
                                    total = pgr_max,
@@ -260,18 +284,27 @@ create_deberta_v2_model<-function(
     message(paste(date(),
               "Creating Tokenizer"))
   }
+  #tokenizer=transformers$PreTrainedTokenizerFast(
+  #  tokenizer_object=tok_new,
+  #  bos_token = "[CLS]",
+  #  eos_token = "[SEP]",
+  #  sep_token = "[SEP]",
+  #  cls_token = "[CLS]",
+  #  unk_token = "[UNK]",
+  #  pad_token = "[PAD]",
+  #  mask_token = "[MASK]",
+  #  add_prefix_space = add_prefix_space,
+  #  trim_offsets=trim_offsets)
+
   tokenizer=transformers$PreTrainedTokenizerFast(
     tokenizer_object=tok_new,
+    unk_token="[UNK]",
+    sep_token="[SEP]",
+    pad_token="[PAD]",
+    cls_token="[CLS]",
+    mask_token="[MASK]",
     bos_token = "[CLS]",
-    eos_token = "[SEP]",
-    sep_token = "[SEP]",
-    cls_token = "[CLS]",
-    unk_token = "[UNK]",
-    pad_token = "[PAD]",
-    mask_token = "[MASK]",
-    add_prefix_space = add_prefix_space,
-    trim_offsets=trim_offsets)
-
+    eos_token = "[SEP]")
 
   if(trace==TRUE){
     message(paste(date(),
@@ -392,6 +425,8 @@ create_deberta_v2_model<-function(
 #'model is stored.
 #'@param raw_texts \code{vector} containing the raw texts for training.
 #'@param p_mask \code{double} Ratio determining the number of words/tokens for masking.
+#'@param whole_word \code{bool} \code{TRUE} if whole word masking should be applied.
+#'If \code{FALSE} token masking is used.
 #'@param val_size \code{double} Ratio determining the amount of token chunks used for
 #'validation.
 #'@param n_epoch \code{int} Number of epochs for training.
@@ -456,6 +491,7 @@ train_tune_deberta_v2_model=function(ml_framework=aifeducation_config$get_framew
                                model_dir_path,
                                raw_texts,
                                p_mask=0.15,
+                               whole_word=TRUE,
                                val_size=0.1,
                                n_epoch=1,
                                batch_size=12,
@@ -643,7 +679,7 @@ train_tune_deberta_v2_model=function(ml_framework=aifeducation_config$get_framew
       return_offsets_mapping = FALSE,
       return_attention_mask = TRUE,
       return_tensors="np",
-      request_word_ids=FALSE,
+      request_word_ids=whole_word,
       report_to_aifeducation_studio=shiny_app_active)),
     remove_columns=raw_text_dataset$column_names
   )
@@ -685,15 +721,26 @@ train_tune_deberta_v2_model=function(ml_framework=aifeducation_config$get_framew
   }
 
   if(ml_framework=="tensorflow"){
-    if(trace==TRUE){
-      message(paste(date(),"Using Token Masking"))
+    if(whole_word==TRUE){
+      if(trace==TRUE){
+        message(paste(date(),"Using Whole Word Masking"))
+      }
+      data_collator=transformers$DataCollatorForWholeWordMask(
+        tokenizer = tokenizer,
+        mlm = TRUE,
+        mlm_probability = p_mask,
+        return_tensors = "tf")
+    } else {
+      if(trace==TRUE){
+        message(paste(date(),"Using Token Masking"))
+      }
+      data_collator=transformers$DataCollatorForLanguageModeling(
+        tokenizer = tokenizer,
+        mlm = TRUE,
+        mlm_probability = p_mask,
+        return_tensors = "tf"
+      )
     }
-
-    data_collator=transformers$DataCollatorForLanguageModeling(
-      tokenizer = tokenizer,
-      mlm = TRUE,
-      mlm_probability = p_mask,
-      return_tensors = "tf")
     tokenized_dataset$set_format(type="tensorflow")
     tokenized_dataset=tokenized_dataset$train_test_split(test_size=val_size)
 
@@ -764,18 +811,28 @@ train_tune_deberta_v2_model=function(ml_framework=aifeducation_config$get_framew
     mlm_model$load_weights(paste0(output_dir,"/checkpoints/best_weights.h5"))
   } else {
 
-    if(trace==TRUE){
-      message(paste(date(),"Using Token Masking"))
+    if(whole_word==TRUE){
+      if(trace==TRUE){
+        message(paste(date(),"Using Whole Word Masking"))
+      }
+      data_collator=transformers$DataCollatorForWholeWordMask(
+        tokenizer = tokenizer,
+        mlm = TRUE,
+        mlm_probability = p_mask,
+        return_tensors = "pt")
+    } else {
+      if(trace==TRUE){
+        message(paste(date(),"Using Token Masking"))
+      }
+      data_collator=transformers$DataCollatorForLanguageModeling(
+        tokenizer = tokenizer,
+        mlm = TRUE,
+        mlm_probability = p_mask,
+        return_tensors = "pt"
+      )
     }
-    data_collator=transformers$DataCollatorForLanguageModeling(
-      tokenizer = tokenizer,
-      mlm = TRUE,
-      mlm_probability = p_mask,
-      return_tensors = "pt")
     tokenized_dataset$set_format(type="torch")
-
     tokenized_dataset=tokenized_dataset$train_test_split(test_size=val_size)
-
 
     if(trace==TRUE){
       message(paste(date(),"Preparing Training of the Model"))